Aggiunto speedup + fix grafico risposta 3

+++ esercizio5/src/it/homepc/tagliabuemichele/model/algorithms/IterationData.java
@@ -3,9 +3,9 @@ package it.homepc.tagliabuemichele.model.algorithms;
 public class IterationData {
     public int iteration;
     public int algorithm;
-    public long elapsedTime;
+    public double elapsedTime;
 
-    public IterationData(int iteration, int algorithm, long elapsedTime) {
+    public IterationData(int iteration, int algorithm, double elapsedTime) {
         this.iteration = iteration;
         this.algorithm = algorithm;
         this.elapsedTime = elapsedTime;
